+++ pkg/k8sclient/awsclusterclient.go
@@ -23,7 +23,7 @@ func NewAWSClusterClient(client client.Client) *AWSClusterClient {
 	}
 }
 
-func (a *AWSClusterClient) Get(ctx context.Context, namespacedName types.NamespacedName) (*capa.AWSCluster, error) {
+func (a *AWSClusterClient) GetAWSCluster(ctx context.Context, namespacedName types.NamespacedName) (*capa.AWSCluster, error) {
 	awsCluster := &capa.AWSCluster{}
 	err := a.client.Get(ctx, namespacedName, awsCluster)
 	if err != nil {
@@ -31,6 +31,16 @@ func (a *AWSClusterClient) Get(ctx context.Context, namespacedName types.Namespa
 	}
 
 	return awsCluster, errors.WithStack(err)
+}
+
+func (a *AWSClusterClient) GetCluster(ctx context.Context, namespacedName types.NamespacedName) (*capi.Cluster, error) {
+	cluster := &capi.Cluster{}
+	err := a.client.Get(ctx, namespacedName, cluster)
+	if err != nil {
+		return nil, errors.WithStack(err)
+	}
+
+	return cluster, errors.WithStack(err)
 }
 
 func (a *AWSClusterClient) GetOwner(ctx context.Context, awsCluster *capa.AWSCluster) (*capi.Cluster, error) {
@@ -72,4 +82,18 @@ func (a *AWSClusterClient) MarkConditionTrue(ctx context.Context, awsCluster *ca
 	originalCluster := awsCluster.DeepCopy()
 	conditions.MarkTrue(awsCluster, condition)
 	return a.client.Status().Patch(ctx, awsCluster, client.MergeFrom(originalCluster))
+}
+
+func (a *AWSClusterClient) Unpause(ctx context.Context, awsCluster *capa.AWSCluster, cluster *capi.Cluster) error {
+	originalCluster := cluster.DeepCopy()
+	cluster.Spec.Paused = false
+	delete(cluster.Annotations, capi.PausedAnnotation)
+	err := a.client.Patch(ctx, cluster, client.MergeFrom(originalCluster))
+	if err != nil {
+		return errors.WithStack(err)
+	}
+
+	originalAwsCluster := awsCluster.DeepCopy()
+	delete(awsCluster.Annotations, capi.PausedAnnotation)
+	return a.client.Patch(ctx, awsCluster, client.MergeFrom(originalAwsCluster))
 }