+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add condition to `AWSCluster` signaling that resolver rules in AWS account have been associated with workload cluster VPC.
+- Add new reconciler that unpauses the `AWSCluster` CR when VPC, Subnets and Resolver Rules conditions are marked as ready.
 
 ## [0.3.0] - 2023-02-22